Undone yet, need several adjusts!

+++ gameserver/realm/entity/player/Player.Init.cs
@@ -541,13 +541,16 @@ namespace gameserver.realm.entity.player
         }
 
         internal Projectile PlayerShootProjectile(
-            byte id, ProjectileDesc desc, ushort objType,
-            int time, Position position, float angle)
+            byte id,
+            ProjectileDesc desc,
+            ushort objType,
+            long time,
+            Position position,
+            float angle
+            )
         {
             ProjectileId = id;
-            return CreateProjectile(desc, objType,
-                (int)StatsManager.GetAttackDamage(desc.MinDamage, desc.MaxDamage),
-                time, position, angle);
+            return CreateProjectile(desc, objType, (int)StatsManager.GetAttackDamage(desc.MinDamage, desc.MaxDamage), time, position, angle);
         }
 
         public override void Tick(RealmTime time)
